+++ javascript/node/selenium-webdriver/index.js
@@ -20,9 +20,10 @@
  * public API and provides convenience assessors to certain sub-modules.
  */
 
-var base = require('./lib/_base');
 var builder = require('./builder');
 var error = require('./error');
+var base = require('./lib/_base');
+var command = require('./lib/command');
 
 
 // NOTE: the remainder of this file is nasty and verbose, but the annotations
@@ -43,10 +44,6 @@ exports.By = require('./lib/by').By;
 
 /** @type {function(new: webdriver.Capabilities)} */
 exports.Capabilities = base.require('webdriver.Capabilities');
-
-
-/** @type {function(new: webdriver.Command)} */
-exports.Command = base.require('webdriver.Command');
 
 
 /** @type {function(new: webdriver.EventEmitter)} */
@@ -96,12 +93,6 @@ exports.WebElementPromise = base.require('webdriver.WebElementPromise');
 /** @type {webdriver.Capability.} */
 (exports.__defineGetter__('Capability', function() {
   return base.require('webdriver.Capability');
-}));
-
-
-/** @type {webdriver.CommandName.} */
-(exports.__defineGetter__('CommandName', function() {
-  return base.require('webdriver.CommandName');
 }));